wgnlpy/nlas: add key + use it
Move all key logic into this NLA.  On decode, 32 zeros is the same as None.
To frob for encode, base64 decode strings to bytes.  It does make sense to
distinguish between None and 32 zeros, since the latter means remove the
private or preshared key.  For public keys, the peer will just not match.

Q: Does it make sense to just move the frob to encode?