wgnlpy/wireguard: replace key frobbing with x25519 wrappers
This allows one to pass in or compare against an explicit key, a key in
base64, or a key as bytes.  The explicit keys are of specific types, which
may prevent confusion between private/public/preshared keys.  It might make
sense to differentiate at the NLA level, but the current scheme seems to
work well enough.