wgnlpy/public_key: add ORCHID-ish address generator
This method converts a PublicKey into an IP address, by default in the
ORCHIDv2 range.  The algorithm concatenates a secret with the key, takes the
shake128 digest equal to the length of the network address, and maps the
digest into the network host range.

The default secret is no secret.. caveat emptor.

This is in contrast to RFC7343, which encodes the hashing algorithm into the
resultant address; since we're only using one algorithm here, we can punt on
avoiding a bidding down attack.  The network prefix also does not affect the
final hash; if that is the desired outcome, said prefix can be mixed in with
the secret.

Note that IPv6 and IPv4 networks do not pull the same number of bits out of
the hash digest, which means that the same public key will generate wildly
different host bits for a regular IPv4 range vs. the same range mapped into
IPv6.  This tradeoff seems acceptable.