wgnlpy/public_key: add LLA-ish address generator
This method converts a PublicKey into an IP address, by default in the Link
Local Address range.  It is similar to the ORCHIDv2 generator, but based on
a truncated Blake2b-256 hash instead of SHAKE-128, as described in [1].

Like the ORCHID generator, it can also take an optional secret, which is
passed as the KEY parameter to the hash.  As before,

	The default secret is no secret.. caveat emptor.

but in this case "no secret" is required for compatibility with wg-lla.  The
only provided test vector is

	PublicKey().lla() == IPv6Address("fe8b:5ea9:9e65:3bc2:b593:db41:30d1:0a4e")

Any IPv6 or IPv4 network can be used to generate the address, but does not
affect the final hash with the exception that the two address families do
not pull the same number of bytes from the front of the final digest.

[1] https://github.com/reidrankin/wg-lla (CC0-1.0)