nlas/sockaddr: fix up value decoding
At least in pyroute2 v0.5.12, using 's' caused NLA to encode and decode an
entire sequence of bytes.  Some time between then and v0.6.4, that went away
and it now decodes a 1-byte sequence.  Normally it would make sense to just
put a fixed length into the format string, but sockaddr is variable length.
Instead, first do the regular decode to get the type & length, then extract
the entire byte sequence as the value.  Note that in the data, there is a
4-byte header we skip.

Q.v. https://github.com/ArgosyLabs/wgnlpy/issues/4